refactor: page.tsx 에서 컴포넌트 분리 관리
