+++ src/components/header/PrintHeader.tsx
@@ -4,7 +4,7 @@ import BackButton from "../common/button/BackButton";
 import { usePathname, useSearchParams } from "next/navigation";
 import { checkPathname } from "@/lib/util/CheckPathName";
 import CustomDialog from "../common/CustomDialog";
-import FontSizeSlider from "../mailDetail/FontSizeSlider";
+import FontSizeSlider from "../mailDetail/mailInfo/FontSizeSlider";
 import { Dialog, DialogTrigger } from "../ui/dialog";
 import { postEpsonPrint, putDeviceSetting } from "@/lib/api/api";
 import Link from "next/link";
